# openapi.yaml
openapi: 3.0.0
paths:
...
  /tenant:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a1
        - TENANT
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a1
  /tenant/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a1
  /tenant/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a1
  /tenant/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a1
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a1
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a1
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a1
  /org:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a2
        - ORG
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a2
  /org/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a2
  /org/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a2
  /org/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a2
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a2
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a2
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a2
  /branch:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a3
        - BRANCH
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branch"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a3
  /branch/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a3
  /branch/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a3
  /branch/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a3
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branch"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a3
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branch"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a3
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a3
  /perm:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a4
        - PERM
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Permission"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a4
  /perm/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a4
  /perm/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a4
...
  /perm/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a4
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Permission"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a4
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Permission"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a4
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a4
  /autoinc:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a5
        - AUTOINC
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Autoincreament"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autoincreament"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a5
  /autoinc/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a5
  /autoinc/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Autoincreament"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a5
...
  /autoinc/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autoincreament"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a5
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Autoincreament"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a5
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Autoincreament"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a5
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autoincreament"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a5
  /docno:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a6
        - DOCNO
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Docnoformat"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Docnoformat"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a6
  /docno/autocomplete:
    post:
      operationId: autoComplete
      summary: ""
      description: retrieve array of {_id, code, name}
      parameters:
        - name: keyword
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/"
        "500":
          description: Internal error
      tags: *a6
  /docno/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Docnoformat"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a6
...
  /docno/{id}:
    get:
      operationId: runFindOne
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Founds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Docnoformat"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a6
    put:
      operationId: runUpdate
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Docnoformat"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a6
    patch:
      operationId: runPatch
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Docnoformat"
      responses:
        "200":
          description: success
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a6
    delete:
      operationId: runDelete
      summary: ""
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Docnoformat"
        "404":
          description: Document not found
        "500":
          description: Internal error
      tags: *a6
  /cat:
    get:
      operationId: runHello
      summary: ""
      description: Say hello only
      parameters: []
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal error
      tags: &a7
        - CAT
    post:
      operationId: runCreate
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "201":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a7
...
  /cat/search:
    post:
      operationId: runSearch
      summary: ""
      parameters: []
      requestBody:
        required: true
        description: Data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiSearchBody"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "400":
          description: bad request
        "500":
          description: internal error
      tags: *a7
...
info:
  title: SimpleApp Demo1
  description: Try CRUD
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-org:
      type: apiKey
      in: header
      name: x-org
      description: "base 64 url encode. example: MC0wLTA/MS0xLTE"
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: /protocol/openid-connect/auth
          scopes: {}
  schemas:
    Object:
      type: object
      properties: {}
    TenantOwner:
      type: object
      properties:
        _id:
          type: string
          default: ""
        label:
          type: string
          default: ""
        uid:
          type: string
          default: ""
      required:
        - _id
        - label
        - uid
    Tenant:
      type: object
      properties:
        _id:
          type: string
          format: uuid
          default: ""
        created:
          type: string
          default: ""
        updated:
          type: string
          default: ""
        createdBy:
          type: string
          default: ""
        updatedBy:
          type: string
          default: ""
        tenantId:
          type: number
          default: 1
        orgId:
          type: number
          default: 1
        branchId:
          type: number
          default: 1
        tenantName:
          type: string
          default: ""
        active:
          type: boolean
          examples:
            - true
          default: true
        description:
          type: string
          default: ""
        owner:
          $ref: "#/components/schemas/TenantOwner"
      required:
        - tenantId
        - tenantName
        - owner
...
security:
  - x-org: []
  - oauth2: []